+++ src/stamina/builder/StaminaModelBuilder.h
@@ -27,6 +27,8 @@
 #include "../util/StateIndexArray.h"
 #include "../util/StateMemoryPool.h"
 
+#include "ProbabilityState.h"
+
 #include <boost/functional/hash.hpp>
 #include <boost/container/flat_map.hpp>
 #include <boost/variant.hpp>
@@ -40,11 +42,6 @@
 namespace stamina {
 	namespace builder {
 
-		// Forward-declare thread classes
-		class BaseThread;
-		class BookKeeperThread : public BaseThread;
-		class ExplorationThread : public BaseThread;
-
 		using namespace storm::builder;
 		using namespace storm::utility::prism;
 		using namespace storm::generator;
@@ -52,104 +49,15 @@ namespace stamina {
 		typedef storm::models::sparse::Ctmc<double> Ctmc;
 		typedef storm::modelchecker::SparseCtmcCslModelChecker<Ctmc> CtmcModelChecker;
 
+
 		template<typename ValueType, typename RewardModelType = storm::models::sparse::StandardRewardModel<ValueType>, typename StateType = uint32_t>
 		class StaminaModelBuilder {
 		public:
-// 			struct
-			/* Sub-class for states with probabilities */
-			class ProbabilityState {
-			public:
-				StateType index;
-				bool assignedInRemapping;
-				uint8_t iterationLastSeen;
-				bool isNew;
-				bool wasPutInTerminalQueue;
-// 				ProbabilityState() : { /* Intentionally left empty */ }
-				ProbabilityState(
-					StateType index = 0
-					, double pi = 0.0
-					, bool terminal = true
-					, uint8_t iterationLastSeen = 0
-				) : index(index)
-					, pi(pi)
-					, terminal(terminal)
-					, assignedInRemapping(false)
-					, iterationLastSeen(iterationLastSeen)
-					, isNew(true)
-					, wasPutInTerminalQueue(false)
-				{
-					// Intentionally left empty
-				}
-				// Copy constructor
-				ProbabilityState(const ProbabilityState & other)
-					: index(other.index)
-					, pi(other.pi)
-					, terminal(other.terminal)
-					, assignedInRemapping(other.assignedInRemapping)
-					, isNew(other.isNew)
-				{
-					// Intentionally left empty
-				}
-
-				double getPi() {
-					return pi;
-				}
-				void addToPi(double add) {
-					pi += add;
-				}
-				void setPi(double pi) {
-					this->pi = pi;
-				}
-				bool isTerminal() {
-					return terminal;
-				}
-				void setTerminal(bool term) {
-					terminal = term;
-				}
-				inline bool operator==(const ProbabilityState & rhs) const {
-					return index == rhs.index;
-				}
-				inline bool operator>=(const ProbabilityState & rhs) const {
-					return index >= rhs.index;
-				}
-				inline bool operator<=(const ProbabilityState & rhs) const {
-					return index <= rhs.index;
-				}
-				inline bool operator>(const ProbabilityState & rhs) const {
-					return index > rhs.index;
-				}
-				inline bool operator<(const ProbabilityState & rhs) const {
-					return index < rhs.index;
-				}
-				double pi;
-				bool terminal;
-
-			};
-			struct ProbabilityStateComparison {
-				bool operator() (
-					const ProbabilityState * first
-					, const ProbabilityState * second
-				) const {
-					// Create a max heap on the reachability probability
-					return first->pi < second->pi;
-				}
-			};
-
-			class ProbabilityStatePair {
-			public:
-				ProbabilityState * first;
-				CompressedState second;
-			};
-
-			struct ProbabilityStatePairComparison {
-				bool operator() (
-					const ProbabilityStatePair first
-					, const ProbabilityStatePair second
-				) const {
-					// Create a max heap on the reachability probability
-					return first.first->pi < second.first->pi;
-				}
-			};
+			// Probability state typedefs
+			typedef ProbabilityState<StateType> ProbabilityState;
+			typedef ProbabilityStateComparison<StateType> ProbabilityStateComparison;
+			typedef ProbabilityStatePair<StateType> ProbabilityStatePair;
+			typedef ProbabilityStatePairComparison<StateType> ProbabilityStatePairComparison;
 
 			/**
 			 * A basic struct for out of order transitions to insert into the transition matrix.
@@ -332,12 +240,9 @@ namespace stamina {
 			storm::expressions::Expression * propertyExpression;
 			storm::expressions::ExpressionManager * expressionManager;
 			std::shared_ptr<const storm::logic::Formula> propertyFormula;
-			storm::storage::sparse::StateStorage<StateType>& stateStorage;
 			std::shared_ptr<storm::generator::PrismNextStateGenerator<ValueType, StateType>> generator;
 			util::StateMemoryPool<ProbabilityState> memoryPool;
-			// StatePriorityQueue statesToExplore;
 			std::deque<std::pair<ProbabilityState *, CompressedState> > statesToExplore;
-			boost::optional<std::vector<uint_fast64_t>> stateRemapping;
 			util::StateIndexArray<StateType, ProbabilityState> stateMap;
 			// Transitions which we must add
 			std::vector<std::vector<TransitionInfo>> transitionsToAdd;
